services/sync-elastic/pyproject.toml
[tool.poetry]
name = "sync-elastic"
version = "0.1.0"
description = ""
authors = ["razboy <mishaneiman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
elasticsearch = "^8.16.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
